% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camelCase.R
\name{camelCase}
\alias{camelCase}
\title{Converts text with separators to camel case (i.e., camelCase)}
\usage{
camelCase(x, sep = "_")
}
\arguments{
\item{x}{Vector of text to be converted}

\item{sep}{The symbol that separates words in the text}
}
\description{
Converts text with separators to camel case (i.e., camelCase)
}
\details{
Eliminates separating character and capitalizes the first letter of each word except the first word. For example converts "This is an example" to thisIsAnExample.
}
