% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catPropPlot.R
\name{catPropPlot}
\alias{catPropPlot}
\title{Creates a color-coded proportion plot good for variables that can be sequentially cumulative (e.g., substrate bigger than x, bigger than x+x1,...)}
\usage{
catPropPlot(cat, prop, time, catOrder, col = gray(seq(0, 1, length.out =
  length(catOrder))), ...)
}
\arguments{
\item{cat}{Categorical data to be plotted (categories for y-variable)}

\item{prop}{Proprortional data (y-variable)}

\item{time}{Time variable (x-variable)}

\item{catOrder}{Order of the categorical variables as a vector in the correct order}

\item{col}{Colors for categories. Should be the same length as length(cat). Defaults to a gray-scale.}
}
\value{
A plot of proportional data over time
}
\description{
Creates a color-coded proportion plot good for variables that can be sequentially cumulative (e.g., substrate bigger than x, bigger than x+x1,...)
}
