% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.bar.r
\name{error.band}
\alias{error.band}
\title{Plots error as a polygon on a base R plot, maybe, I haven't really tested or used this much}
\usage{
error.band(x, y, upper.y, lower.y = upper.y, col = "gray", border = NA,
  interval.type = "dist", ...)
}
\arguments{
\item{x}{a numeric vector palcement of error bars on the x-axis}

\item{y}{a numeric vector of the y-origins of the error bars}

\item{upper.y}{A vector of the upper y limits for error bars}

\item{lower.y}{A vector of the lower y limits for error bars, set to upper.y by default}

\item{col}{Color of the polygon}

\item{border}{Color for the border of the polygon}

\item{interval.type}{If this is set to "dist", upper.y is set to y+upper.y}
}
\description{
Plots error as a polygon on a base R plot, maybe, I haven't really tested or used this much
}
\details{
Puts error bars on a plot. The default has only y error bars, you must define the vectors x,y, and upper.y (your error vector) at a minimum. Lower and upper error can be set separately by also defining lower.y. To display x error bars set x.bar=T and upper.x to a vector containing your x error. type is set to either "dist" as in distance from the mean or "value" which is the actual lower/upper bound
}

