% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutliHist.R
\name{multiHist}
\alias{multiHist}
\title{Creates a stacked histogram (as opposed to side-by-side or overlapping) for categorical data}
\usage{
multiHist(data, y, group, breaks, col = NULL, axisStep = 1, xlimit = NULL,
  logAxis = F, axisLim = NULL, ...)
}
\arguments{
\item{data}{is a data.frame or data.table containing the data}

\item{y}{is the variable to be plotted as a histogram as a quoted name}

\item{group}{is the quoted name of the grouping variable}

\item{breaks}{is a vector of breaks for the histogram}

\item{col}{is an optional specification of the colors to be used in order of unique(group)}

\item{axisStep}{is the spacing of x-axis labels}
}
\value{
A stacked histogram color coded by the provided factor
}
\description{
Creates a stacked histogram (as opposed to side-by-side or overlapping) for categorical data
}

