% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelLabel.R
\name{panelLabel}
\alias{panelLabel}
\title{Puts text in the top left corner of a plot}
\usage{
panelLabel(text, xadj = 0.035, yadj = 0.05)
}
\arguments{
\item{text}{the text to print}

\item{xadj}{position along the x-axis as a proportion of the axis span}

\item{yadj}{position along the y-axis as a proportion of the axis span}

\item{position}{currently only topLeft}
}
\description{
Puts text in the top left corner of a plot
}

