% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFiles.R
\name{png.par}
\alias{png.par}
\title{Wrapper functions for fast creation of plot files with defaults I like
Just combines the grahic device functions png, tiff, and pdf with par to open a file for plotting}
\usage{
png.par(file, res = 200, width = 6.5, height = 5, units = "in",
  bg = NA, mar = c(2.5, 2.5, 0, 0), mgp = c(1.5, 0.5, 0), las = 1,
  bty = "l", col = "black", ...)
}
\arguments{
\item{file}{The file to write to, this is the only necessary argument}
}
\description{
Wrapper functions for fast creation of plot files with defaults I like
Just combines the grahic device functions png, tiff, and pdf with par to open a file for plotting
}
\details{
Any argument from the respective grahic function or par can be passed, but the only thing that is strictly necessary is the file path, others are set with defaults I like.
}
